feat: Update dependencies in hardhat and react-app packages
Update the dependencies in the hardhat and react-app packages to their latest versions. This includes updating the hardhat-ethers package to version 1.0.0 in both packages. Also, update the @wagmi/connectors package to version 5.1.11 in the react-app package. Additionally, update the ethers package to version 6.13.2 and the next package to version 14.2.13 in the react-app package.